requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
hydra-core>=1.3.0
hydra-colorlog>=1.2.0
omegaconf>=2.3.0
wandb>=0.15.0
tensorboard>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
jiwer>=3.0.0
nltk>=3.8.0
